pax: Stricter CRS handling
Assume a default CRS (which we need to use H3 anyway), and convert data
on the way in.

Use geom_types as our flag variable, which we'll need later to decide
what we're doing with incoming data.

Raises even more questions with the new_strata shapefiles, but save
those for another day.